# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
...
  models.Contact:
    description: Company Contacts
    properties:
      email:
        description: Email
        type: string
      firstName:
        description: First name
        type: string
      info:
        description: Info
        type: string
      lastName:
        description: Last name
        type: string
      phone:
        description: Phone/cel
        type: string
    type: object
  models.Framework:
    properties:
      description:
        description: Description
        type: string
      id:
        type: string
      name:
        description: Framework name
        type: string
      phases:
        description: Pentest phases
        items:
          $ref: '#/definitions/models.Phase'
        type: array
    type: object
  models.Phase:
    properties:
      description:
        description: Description
        type: string
      info:
        description: Information
        type: string
      name:
        description: Phase name
        type: string
      status:
        description: |-
          Status of completion
          example: TODO|DOING|DONE
        type: string
      tools:
        description: Phase common used tools
        items:
          $ref: '#/definitions/models.Tool'
        type: array
      url:
        description: URL
        type: string
    type: object
  models.Project:
    properties:
      companyId:
        description: Company ID
        type: string
      createdAt:
        description: Create Time
        type: string
      description:
        description: Description
        type: string
      framework:
        allOf:
        - $ref: '#/definitions/models.Framework'
        description: Framework to use for the project
      id:
        type: string
      name:
        description: Project Name
        type: string
      status:
        description: 'Status: Approval|Open|Executing|Reporting|Retesting|Done'
        type: string
      targets:
        description: 'Targets: URL and/or IP'
        items:
          $ref: '#/definitions/models.Target'
        type: array
      type:
        description: 'Project type: BlackHat|GrayHat|WhiteHat'
        type: string
    type: object
  models.Target:
    properties:
      description:
        description: Description
        type: string
      ip:
        description: Ip address
        type: string
      name:
        description: Target name
        type: string
      network:
        description: Network/AS
        type: string
      owner:
        description: Whois Owner
        type: string
      url:
        description: URL
        type: string
    type: object
  models.Tool:
    properties:
      description:
        description: Description
        type: string
      name:
        description: Tool name
        type: string
      synopsis:
        description: Synopsis and examples
        type: string
      url:
        description: URL for reference
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: mentesan@gmail.com
    name: Fabio Almeida
    url: https://github.com/mentesan
  license:
    name: MIT
    url: https://github.com/mentesan/pert-api/blob/main/LICENSE
  title: Pentest Execution and Reporting Tool (PERT) API
  version: 0.0.1
paths:
...
  /contacts:
    delete:
      parameters:
      - description: Contact.ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete contact
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get contacts list
    post:
      parameters:
      - description: JSON for new contact
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Add new contact
    put:
      parameters:
      - description: Contact.ID
        in: query
        name: id
        required: true
        type: string
      - description: All fields are optional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Contact'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Update contact information
  /contacts/search:
    get:
      parameters:
      - description: For search
        in: body
        name: request
        required: true
        schema:
          example: '{ "name": "ContactFirstName" }'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Search contact by FirstName
  /projects:
    delete:
      parameters:
      - description: Project.ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Delete project
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get projects list
    post:
      parameters:
      - description: JSON for new project
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Add new project
    put:
      parameters:
      - description: Project.ID
        in: query
        name: id
        required: true
        type: string
      - description: All fields are optional
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      summary: Update project information
  /projects/search:
    get:
      parameters:
      - description: For search
        in: body
        name: request
        required: true
        schema:
          example: '{ "name": "ProjectName" }'
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Search project by Name
...
produces:
- application/json
schemes:
- http
swagger: "2.0"